+++ qiskit/qobj/__init__.py
@@ -10,6 +10,4 @@
 from .qobj import Qobj
 from .models import (QobjConfig, QobjExperiment, QobjInstruction, QobjHeader,
                      QobjExperimentHeader, QobjConditional, QobjExperimentConfig)
-from .exceptions import QobjValidationError
-
-from ._validation import validate_qobj_against_schema
+from .utils import validate_qobj_against_schema
