+++ qiskit/qobj/utils.py
@@ -9,8 +9,22 @@
 
 from enum import Enum
 
+from qiskit.validation.jsonschema import validate_json_against_schema
+
 
 class QobjType(str, Enum):
     """Qobj.type allowed values."""
     QASM = 'QASM'
     PULSE = 'PULSE'
+
+
+def validate_qobj_against_schema(qobj):
+    """Validates a QObj against the .json schema.
+
+    Args:
+        qobj (Qobj): Qobj to be validated.
+    """
+    validate_json_against_schema(
+        qobj.as_dict(), 'qobj',
+        err_msg='Qobj failed validation. Set Qiskit log level to DEBUG '
+                'for further information.')
